@@ -122,10 +122,10 @@
     "@types/lodash-es": "4.17.9",
     "@types/marked": "5.0.1",
     "@types/node": "18.17.17",
-    "@typescript-eslint/eslint-plugin": "^6.10.0",
-    "@typescript-eslint/parser": "^6.10.0",
+    "@typescript-eslint/eslint-plugin": "6.13.1",
+    "@typescript-eslint/parser": "6.13.1",
     "autoprefixer": "10.4.15",
-    "eslint": "^8.53.0",
+    "eslint": "^8.54.0",
     "eslint-config-prettier": "8.10.0",
     "eslint-plugin-import": "2.28.1",
     "eslint-plugin-jsdoc": "46.8.1",
