@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { RouterStateSnapshot, ActivatedRouteSnapshot } from '@angular/router';
+import { ActivatedRouteSnapshot } from '@angular/router';
 import { Observable } from 'rxjs';
 import { DocumentsService } from './documents.service';
 import { Document } from './models/document';
@@ -10,11 +10,9 @@ import { Document } from './models/document';
 export class DocumentResolver {
   constructor(private documentsService: DocumentsService) {}
 
-  resolve(
-    route: ActivatedRouteSnapshot,
-    state: RouterStateSnapshot,
-  ): Observable<Document> {
-    const documentName = route.data.documentName ?? route.params.documentName;
+  resolve(route: ActivatedRouteSnapshot): Observable<Document> {
+    const documentName =
+      route.data['documentName'] ?? route.params['documentName'];
     return this.documentsService.fetchDocument(documentName);
   }
 }
