@@ -2,6 +2,7 @@ import { Injectable } from '@angular/core';
 import { SocketFactoryService } from './socket-factory.service';
 import { Observable } from 'rxjs';
 
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
 type Listener = (...args: any[]) => void;
 
 @Injectable({
@@ -22,7 +23,7 @@ export class Socket {
     return this;
   }
 
-  call<T>(methodName: string, ...args: any[]): Observable<T> {
+  call<T>(methodName: string, ...args: unknown[]): Observable<T> {
     return new Observable(observer => {
       const sendArgs = args.length > 0 ? args : [{}]; // we need to pass at least one argument due to NestJS limitation
       this.socket.emit(methodName, ...sendArgs, (response: T) => {
