@@ -8,6 +8,6 @@ describe('playerById', () => {
         1: { id: '1', name: 'wonszu' } as Player,
         2: { id: '2', name: 'maly' } as Player,
       }),
-    ).toEqual({ id: '2', name: 'maly' } as any);
+    ).toEqual({ id: '2', name: 'maly' } as Player);
   });
 });
