@@ -1,7 +1,6 @@
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { QueueClassSlotListComponent } from './queue-class-slot-list.component';
 import { provideMockStore, MockStore } from '@ngrx/store/testing';
-import { MemoizedSelector } from '@ngrx/store';
 import { mySlot } from '../queue.selectors';
 import { MockComponent } from 'ng-mocks';
 import { QueueSlotContainerComponent } from '../queue-slot-container/queue-slot-container.component';
@@ -14,7 +13,6 @@ describe('QueueClassSlotListComponent', () => {
   let component: QueueClassSlotListComponent;
   let fixture: ComponentFixture<QueueClassSlotListComponent>;
   let store: MockStore;
-  let mySlotSelector: MemoizedSelector<unknown, any>;
 
   const initialState = {
     queue: {
@@ -53,7 +51,7 @@ describe('QueueClassSlotListComponent', () => {
 
   beforeEach(() => {
     store = TestBed.inject(MockStore);
-    mySlotSelector = store.overrideSelector(mySlot, null);
+    store.overrideSelector(mySlot, null);
     fixture = TestBed.createComponent(QueueClassSlotListComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
